pyproject.toml
[tool.poetry]
name = "coursework-5"
version = "0.1.0"
description = ""
authors = ["alexandr <dmitrymaslov2016@mai.ru>"]
readme = "README.md"
packages = [{include = "coursework_5"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.29.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
